% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nephoshapes.R
\name{classifyModel}
\alias{classifyModel}
\title{Classify clouds in a model}
\usage{
classifyModel(mdata, mname, ttmx_dir, suffix = ".ttmx.dist.pac")
}
\arguments{
\item{mdata}{Data of a model, as given by \code{\link{summarizeHDBSCAN}}.}

\item{mname}{Name of a model; the name of \code{mdata}.}

\item{ttmx_dir}{Directory where the token-level distance matrices are stored.}

\item{suffix}{Suffix of the file names for the token-level distance matrices;
the function assumes that the name of the file is the name of the medoid
plus the suffix.}
}
\value{
A table with one row per cluster in the model, the columns created by
\code{\link{clusterSemvar}}, \code{\link{clusterHDBSCAN}} and
\code{\link{clusterDistance}} and the classification of each cluster based
on the Nephological Shapes from \insertCite{montes_2021;textual}{semcloud}
(see \href{https://cloudspotting.marianamontes.me/shapes.html}{Chapter 5}
for a full description and examples).
}
\description{
Classify clouds in a model
}
\examples{
\dontrun{
purrr::imap_dfr(models$medoidCoords, classifyModel, ttmx_dir = 'path/to/dir')
}
}
\references{
\insertAllCited{}
}
