% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concordancer.R
\name{cleanWord}
\alias{cleanWord}
\title{Clean Word column}
\usage{
cleanWord(txt)
}
\arguments{
\item{txt}{word form}
}
\value{
Converted string.
}
\description{
Function to convert word forms coming from the \emph{QLVLNewsCorpus} towards
something more friendly for HTML rendering.
}
