% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nephoshapes.R
\name{clusterHDBSCAN}
\alias{clusterHDBSCAN}
\title{Summarize HDBSCAN data per cluster}
\usage{
clusterHDBSCAN(m)
}
\arguments{
\item{m}{Tibble with one token per row and HDBSCAN information. The \code{coords}
element of a model resulting from \code{\link{summarizeHDBSCAN}}.}
}
\value{
Tibble with one row per cluster and various HDBSCAN-derived values:
\describe{
\item{min_, mean_, max_ and sd_cws}{Minimum, mean and maximum, as well
as standard deviation, of the number of first-order context words per
token in that cluster.}
\item{min_, mean_, max_ and sd_eps}{Minimum, mean and maximum, as well
as standard deviation, of the \eqn{\epsilon} value of the tokens in
that cluster.}
\item{size, rel_size}{Absolute number of tokens in the cluster and proportion
of \emph{modelled} tokens covered by the cluster.}
\item{deeper_than_noise}{Proportion of tokens in that cluster with an
\eqn{\epsilon} value lower than the minimum \eqn{\epsilon} of noise
tokens in that model.}
\item{cw_tokens, _types, _ttratio}{Union of first-order context words
of tokens in that cluster: number of types and of tokens and type-token
ratio.}
}
}
\description{
Summarize HDBSCAN data per cluster
}
