% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nephoshapes.R
\name{clusterSemvar}
\alias{clusterSemvar}
\title{Compute Semvar values on clusters}
\usage{
clusterSemvar(m)
}
\arguments{
\item{m}{Tibble with one token per row and HDBSCAN information. The \code{coords}
element of a model resulting from \code{\link{summarizeHDBSCAN}}.}
}
\value{
A tibble with one row per cluster and output from
\code{\link[semvar]{clusterqualkNN}} and \code{\link[semvar]{clusterqualSIL}}
for each class, based on the \emph{coordinates} in the input and both including
and excluding noise tokens.
}
\description{
Compute Semvar values on clusters
}
