% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concordancer.R
\name{correctDistance}
\alias{correctDistance}
\title{Correct Distance values}
\usage{
correctDistance(cws, distance_corrector_func)
}
\arguments{
\item{cws}{Dataframe as outputted by \code{listContextwords}
(\href{https://github.com/montesmariana/semasioFlow}{semasioFlow} Python module).}

\item{distance_corrector_func}{Function to be applied to the \code{word} variable to remove irrelevant lines.}
}
\value{
Tibble with the context word ids and the corrected distance as \code{conc_distance}.
}
\description{
Remove rows that should not count in the distance computation and recalculate distances.
}
