% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareLemmas.R
\name{customDist}
\alias{customDist}
\title{Compute distances between models}
\usage{
customDist(
  mnames,
  input_dir,
  transformed = TRUE,
  fun = c("euclidean", "procrustes", "mantel")
)
}
\arguments{
\item{mnames}{List of model names.}

\item{input_dir}{Directory where the models are stored.}

\item{transformed}{Whether the distance matrices}

\item{fun}{Function to calculate the distances.}
}
\value{
A distance matrix (\code{matrix} object) with models as rows and columns.
}
\description{
Three functions to compute distances between (token) matrices are supported:
\itemize{
\item "euclidean" runs \code{\link{eucliMats}}, which computes tokenwise euclidean distances and averages over them;
\item "procrustes" runs \code{\link[vegan]{procrustes}} via \code{\link{procMats}};
\item "mantel" runs \code{\link[vegan]{mantel}} via \code{\link{mantelMats}}.
}
}
