% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdbscan.R
\name{cwsForClusters}
\alias{cwsForClusters}
\title{Map context words and HDBSCAN clusters}
\usage{
cwsForClusters(variables, cws_column, cluster_column, b = 1)
}
\arguments{
\item{variables}{Dataframe with IDs, clusters and lists of context words}

\item{cws_column}{Name of the column with the character vectors (one per row) of context words}

\item{cluster_column}{Name of the column with the name of the clusters (as factor)}

\item{b}{Weight for computing \code{\link{fscore}}}
}
\value{
a \link[tibble:tibble-package]{tibble} with one row per context word per cluster, with frequency information.
}
\description{
Map context words and HDBSCAN clusters
}
