% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{eucliMats}
\alias{eucliMats}
\title{Mean euclidean distance}
\usage{
eucliMats(mat1, mat2)
}
\arguments{
\item{mat1}{First matrix}

\item{mat2}{Second matrix}
}
\value{
Distance between matrix
}
\description{
Computes the euclidean distance between the individual vectors
and takes their mean.
}
\seealso{
Other distances: 
\code{\link{mantelMats}()},
\code{\link{procMats}()}
}
\concept{distances}
