% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concordancer.R
\name{filterWeight}
\alias{filterWeight}
\title{Filter by Weighting Information}
\usage{
filterWeight(
  weight_param,
  tid_data,
  weight_filter_fun = function(weightparam) stringr::str_ends(weightparam, "no", negate
    = TRUE),
  weight_as_sup = FALSE,
  threshold = 0
)
}
\arguments{
\item{weight_param}{Character string coding the relevant weighting information.}

\item{tid_data}{Subsection of a context-word-by-token dataframe enriched by
\code{\link{filterFoc}}.}

\item{weight_filter_fun}{Function that takes \code{weight_param} as input and
returns \code{TRUE} if weighting should be included and \code{FALSE} if it
should be ignored.}

\item{weight_as_sup}{Should weight values be included in superscript?}

\item{threshold}{Weighting threshold for a context word to be included}
}
\value{
Enriched dataframe including columns with filtering information.
}
\description{
Filter by Weighting Information
}
