% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{focdists_from_csv}
\alias{focdists_from_csv}
\title{Open type level distance matrices
Open a distance matrix stored as a \code{.csv} file. Replaces
\code{\link{tokens_from_pac}} for type level matrices.}
\usage{
focdists_from_csv(input_directory, filename)
}
\arguments{
\item{input_directory}{Directory where distance matrices
(as output from python module) are stored}

\item{filename}{Not full name of a distance matrix stored
in \code{input_directory}.}
}
\value{
Distance matrix as matrix with ids as rownames and column names.
}
\description{
Open type level distance matrices
Open a distance matrix stored as a \code{.csv} file. Replaces
\code{\link{tokens_from_pac}} for type level matrices.
}
