% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{fscore}
\alias{fscore}
\title{F-score}
\usage{
fscore(precision, recall, b = 1)
}
\arguments{
\item{precision}{Precision}

\item{recall}{Recall}

\item{b}{Weight}
}
\value{
F-score
}
\description{
Compute F-score from precision and recall.
}
