% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClouds.R
\name{getClouds}
\alias{getClouds}
\title{Get Clouds from distance matrix}
\usage{
getClouds(
  input_dir,
  output_dir,
  files_list,
  lemma,
  solutions,
  logrank = TRUE,
  logdist = TRUE,
  type = "token",
  row_selection = vector()
)
}
\arguments{
\item{input_dir}{Directory where the token distance matrices are stored.}

\item{output_dir}{Directory where the data will be stored.}

\item{files_list}{Liste of filenames within \code{input_dir}.}

\item{lemma}{Name of the lemma, for filenames}

\item{solutions}{Named list of techniques to run for visualization possible \code{technique} values in \code{\link{getFit}}.}

\item{logrank}{Whether to transform the matrices with \code{\link{transformMats}}.}

\item{logdist}{Whether euclidean distances should be computed between the rows
of the transformed matrices (when \code{logrank} is \code{TRUE}). Otherwise,
the matrix of log-transformed ranks is only made symmetric.}

\item{type}{Whether to open the files with \code{\link{tokensFromPac}}
(for "token") or \code{\link{focdistsFromCsv}} (otherwise).}

\item{row_selection}{List of row (and column) names to subset the matrices.}
}
\value{
List of stresses (emtpy if "mds" is not given.)
}
\description{
Read distance matrices from different models,
run dimensional reduction for visualization based on different techniques
and store the coordinates corresponding to each model in a dataframe per technique.
The names of the models will be found in the \code{models} file and their paths will be
searched for in \code{input_dir}: if a file is not found, a warning will be issued.
}
