% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concordancer.R
\name{getContext}
\alias{getContext}
\title{Context from concordance dataframe}
\usage{
getContext(
  tid,
  cws,
  foc_param = NA,
  weight_param = NA,
  cw_selection = NA,
  clean_word_fun = cleanWord,
  to_remove = c("<sentence>"),
  ...
)
}
\arguments{
\item{tid}{Token ID.}

\item{cws}{Dataframe with one row per context word per token, as outputted
by \code{\link{setupConcordancer}}.}

\item{foc_param}{Character string with information on first order filters.}

\item{weight_param}{Character string with information on weighting filters.}

\item{cw_selection}{List or character string of semicolon-separated values
with the context words captured by the model.}

\item{clean_word_fun}{Function to clean the \code{word} column.}

\item{to_remove}{Vector of \code{word} values to remove altogether.}

\item{...}{Arguments to be passed to \code{\link{filterFoc}} and \code{\link{filterWeight}}.}
}
\value{
Character string with (weighted) concordance line.
}
\description{
From a dataframe with one row per context word of a token and columns for
elements (in particular, \code{word} with the appropriate enrichment),
create a concordance line.
}
\examples{
\dontrun{
cws <- setupConcordancer(lemma, input_dir)
getContext(tokenID, cws, 'bound5-5lex', 'PPMIweight', c('a/det', 'horse/noun', 'fast/adj'))
}

}
