% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{getCoords}
\alias{getCoords}
\title{Extract coordinates from fit}
\usage{
getCoords(fit, modelname, rownames, d = "", source = "tsne")
}
\arguments{
\item{fit}{Dimensionality reduction result; output from \code{\link{getFit}}}

\item{modelname}{Name of the model, for the names of the columns}

\item{rownames}{List of names for the rows}

\item{d}{Output from previous run or, if it's the first run, empty string.}

\item{source}{Technique used in dimensionality reduction: either "mds"
for output from \code{\link[vegan]{metaMDS}} or "tsne" for \code{\link[Rtsne]{Rtsne}}
(it matches \code{technique} from \code{\link{getFit}}).}
}
\value{
Tibble with the coordinates of each element.
}
\description{
Extract coordinates from fit
}
