% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{getDistMtx}
\alias{getDistMtx}
\title{Get Distance Matrix}
\usage{
getDistMtx(wwmx, source_dir, lemma, suffix = "")
}
\arguments{
\item{wwmx}{Distance matrix as outputted from \code{\link{focdistsFromCsv}}.}

\item{source_dir}{Directory to store the distance matrix in}

\item{lemma}{Name of the lemma, for the filename}

\item{suffix}{Suffix to add after \verb{[lemma].models.dist} in the filename}
}
\value{
Distance matrix as a \code{dist} object.
}
\description{
Store distances as \code{tsv} and return \code{dist} object.
}
