% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{getFit}
\alias{getFit}
\title{Dimensionality reduction for visualization}
\usage{
getFit(d, dim = 2, technique, perp = 30, seed = 8541)
}
\arguments{
\item{d}{distance matrix, which will be forced into a \code{matrix} object
and then a \code{dist} object}

\item{dim}{number of dimensions, by default 2. As argument \code{k} for
\code{\link[vegan]{metaMDS}} and \code{dims} for \code{\link[Rtsne]{Rtsne}}}

\item{technique}{either "mds" to run \code{\link[vegan]{metaMDS}},
"tsne" to run \code{\link[Rtsne]{Rtsne}} or "umap" to run \code{\link[umap]{umap}}}

\item{perp}{perplexity value for \code{\link[Rtsne]{Rtsne}}, default is 30.
This value is ignored when \code{\link[vegan]{metaMDS}} is run.}

\item{seed}{seed to keep randomness at check}
}
\value{
output of either \code{\link[vegan]{metaMDS}} or \code{\link[Rtsne]{Rtsne}}.
}
\description{
Wrapper around \code{\link[vegan]{metaMDS}} and \code{\link[Rtsne]{Rtsne}}
for dimensionality reduction with certain set parameters.
}
\details{
\code{\link[vegan]{metaMDS}} is run with \code{trymax=20} and \code{trace=FALSE}
by default, so that the search for best solution is not printed.
\code{\link[Rtsne]{Rtsne}} uses as default parameters
\code{theta=0.0, check.duplicates = FALSE, max_iter = 1000}.
}
