% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{matricizeCloud}
\alias{matricizeCloud}
\title{Turn tibble into matrix}
\usage{
matricizeCloud(df, rownames = "X_model")
}
\arguments{
\item{df}{Tibble}

\item{rownames}{Name of the column that will become the row names. Default is
"X_model", as it would work with the distances tibble.}
}
\value{
Matrix with right row names
}
\description{
Turn tibble into matrix
}
