% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-filter.R
\name{posFilter}
\alias{posFilter}
\title{Extract part-of-speech filter}
\usage{
posFilter(foc_param)
}
\arguments{
\item{foc_param}{Character string coding the relevant first-order parameters.}
}
\value{
Character vector: empty if no \code{pos} filter should be applied,
values to be selected otherwise.
}
\description{
Extract part-of-speech filter
}
