% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concordancer.R
\name{reorderPpmi}
\alias{reorderPpmi}
\title{Reorder PPMI columns}
\usage{
reorderPpmi(ppmi, pmi_columnname)
}
\arguments{
\item{ppmi}{Dataframe with a row per context word and columns for frequency
and association measures/}

\item{pmi_columnname}{Name of the column with the weighting values of interest.}
}
\value{
Dataframe with a \code{cw} column for the context word and a
\code{my_weight} column for the weighting.
}
\description{
The values of the \code{ppmi} will be used for weight selection and can be
included as superindices.
}
