% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdbscan.R
\name{summarizeHDBSCAN}
\alias{summarizeHDBSCAN}
\title{Summarize HDBSCAN data for a model}
\usage{
summarizeHDBSCAN(
  lemma,
  modelname,
  input_dir,
  output_dir,
  minPts = 8,
  includePlot = FALSE,
  coords_name = ".tsne.30"
)
}
\arguments{
\item{lemma}{Name of the lemma, for filenames}

\item{modelname}{Name of the model, for coordinates and filename}

\item{input_dir}{Directory where the distance matrix is stored}

\item{output_dir}{Directory where coordinates are stored. This directory must contain:
\itemize{
\item A file with the coordinates of the tokens, with a name combining \code{lemma} and \code{coords_name} and ending in \code{.tsv}.
\item A file with coordinates for the context words, with a name combining \code{lemma} and \code{coords_name} and ending in \code{.cws.tsv}.
\item A file with semicolon-separated lists of context words, with a name combining \code{lemma} and \code{.variables.tsv}
}}

\item{minPts}{Minimum points for \code{\link[dbscan]{hdbscan}}}

\item{includePlot}{Whether too include the plot (requires \code{cowplot}.)}

\item{coords_name}{The code in the coordinate files indicating the type of dimensionality reduction performed, for filenames}
}
\value{
list with at least two items:
\itemize{
\item{\strong{coords}: }{\code{\link[tibble]{tibble}} with one row per token, the coordinates in the pertinent file, and information
from \code{\link{extractHDBSCAN}}} as well as the \code{variables} file.
\item{\strong{cws}: }{\code{\link[tibble]{tibble}} with one row per context word and cluster, output
from \code{\link{cwsForClusters}}}, combined with coordinates from the relevant file.
\item{\strong{hplot}: }{If \code{includePlot}, the HDBSCAN plot.}
}
}
\description{
Summarize HDBSCAN data for a model
}
