% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{tokens_from_pac}
\alias{tokens_from_pac}
\title{Open token distance matrix}
\usage{
tokens_from_pac(input_directory, filename)
}
\arguments{
\item{input_directory}{Directory where distance matrices
(as output from python module) are stored}

\item{filename}{Not full name of a distance matrix stored
in \code{input_directory}.}
}
\value{
Distance matrix as matrix with ids as rownames and column names.
}
\description{
Open a distance matrix stored as a \code{.pac} compressed file
with a \code{.npy} file with distances and a \code{.meta} file
with dimension names. Might only work for token-to-token distance
matrices.
}
