% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{tokensFromPac}
\alias{tokensFromPac}
\title{Open token distance matrix}
\usage{
tokensFromPac(input_file)
}
\arguments{
\item{input_file}{Name of the file}
}
\value{
Distance matrix as matrix with ids as rownames and column names.
}
\description{
Open a distance matrix stored as a \code{.pac} compressed file
with a \code{.npy} file with distances and a \code{.meta} file
with dimension names. Might only work for token-to-token distance
matrices.
}
