% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{transformMats}
\alias{transformMats}
\title{Log-transform distance matrices}
\usage{
transformMats(mat, asDist = TRUE)
}
\arguments{
\item{mat}{Distance matrix to transform}

\item{asDist}{Boolean: whether to force the matrix into a symmetric distance matrix.}
}
\value{
Double logged transformation of ranked rows of the input matrix,
as \code{matrix}.
}
\description{
Log-transform distance matrices
}
