% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concordancer.R
\name{weightConcordance}
\alias{weightConcordance}
\title{Create weighted concordance lines}
\usage{
weightConcordance(variables, cws, lemma, ...)
}
\arguments{
\item{variables}{Dataframe with one row per token ID and at least columns
prefixed by \code{_cws.} with semicolon-separated context words.}

\item{cws}{Dataframe with one row per token ID per context word as outputted
by \code{\link{setupConcordancer}}.}

\item{lemma}{Name of the lemma, to process model names}

\item{...}{Arguments to be passed to \code{\link{weightLemma}} and
\code{\link{getContext}}, in order to adapt to different ways of coding
parameter settings. See \code{vignette('weightConcordance')}.}
}
\value{
Enriched \code{variables} dataframe with columns containing weighted
concordance lines.
}
\description{
This code assumes that the lists of context words are stored in columns
starting with \code{_cws}, followed by the name of the lemma and a model name
from which first-order parameter settings can be extracted (all separated
by periods). Then it creates
the weighted concordance lines and stores them in columns following the same
name pattern, but starting with \code{_ctxt}. In addition, it creates a non
weighted column \code{_ctxt.raw}.
}
\examples{
\dontrun{
cws <- setupConcordancer(lemma, input_dir)
variables <- readr::read_tsv('path/to/file', lazy = F)
ctxts <- weightConcordance(variables, cws, lemma)
}

}
