% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concordancer.R
\name{weightLemma}
\alias{weightLemma}
\title{Tailor weighting function to a lemma}
\usage{
weightLemma(
  cws,
  foc_param_fun = function(m) stringr::str_split(m, "\\\\.")[[1]][[1]],
  weight_param_fun = function(m) stringr::str_split(m, "\\\\.")[[1]][[2]],
  sup_weight_fun = function(weightparam) stringr::str_ends(weightparam, "weight"),
  ...
)
}
\arguments{
\item{cws}{Dataframe with one row per token per context word, output
of \code{\link{setupConcordancer}}.}

\item{foc_param_fun}{Function that takes the name of a model and return a
character string with first-order filters (to be used as \code{foc_param}
in \code{\link{getContext}} and thus \code{\link{filterFoc}}).}

\item{weight_param_fun}{Function that takes the name of a model and return a
character string with weighting filters (to be used as \code{weight_param}
in \code{\link{getContext}} and thus \code{\link{filterWeight}}).}

\item{sup_weight_fun}{Function that takes the output of \code{weight_param_fun}
as input and returns \code{TRUE} if weighting values should be added as
superindices and \code{FALSE} if they should not. To be fed to \code{\link{filterWeight}}.}

\item{...}{Arguments to be passed to \code{\link{getContext}}.}
}
\value{
Function that takes a token ID, a list of context words and a model name
as input and calls \code{\link{getContext}}.
}
\description{
Tailor weighting function to a lemma
}
