% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-filter.R
\name{windowFilter}
\alias{windowFilter}
\title{Extract window filter}
\usage{
windowFilter(foc_param)
}
\arguments{
\item{foc_param}{Character string coding the relevant first-order parameters.}
}
\value{
Integer vector of length 2 (or list) with left and right window span lengths.
}
\description{
Extract window filter
}
