% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nephoshapes.R
\name{clusterSeparation}
\alias{clusterSeparation}
\title{Compute Semvar values on clusters}
\usage{
clusterSeparation(m, k = 8)
}
\arguments{
\item{m}{Tibble with one token per row and HDBSCAN information. The \code{coords}
element of a model resulting from \code{\link{summarizeHDBSCAN}}.}

\item{k}{Number of tokens to compute \code{\link{separationkNN}}}
}
\value{
A tibble with one row per cluster and output from
\code{\link{separationkNN}} and \code{\link[cluster]{silhouette}}
for each class, based on the \emph{coordinates} in the input and both including
and excluding noise tokens.
}
\description{
Compute Semvar values on clusters
}
