% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{focdistsFromCsv}
\alias{focdistsFromCsv}
\title{Open type level distance matrices
Open a distance matrix stored as a \code{.csv} file. Replaces
\code{\link{tokensFromPac}} for type level matrices.}
\usage{
focdistsFromCsv(input_file)
}
\arguments{
\item{input_file}{Name of the file}
}
\value{
Distance matrix as matrix with ids as rownames and column names.
}
\description{
Open type level distance matrices
Open a distance matrix stored as a \code{.csv} file. Replaces
\code{\link{tokensFromPac}} for type level matrices.
}
